Refactor bounds types into dedicated files
